GemmaAgent/requirements.txt
python-dotenv
langchain
langchain_classic
langchain-community
langchain-text-splitters
langchain-astradb
langchain-google-genai
langchain-groq
pypdf
chromadb
tqdm
requests
numpy
pandas
streamlit
huggingface-hub
sentence-transformers
